reorganize and basic-auth
